pyproject.toml
[tool.poetry]
name = "be4-fastapi"
version = "0.1.0"
description = ""
authors = ["taewoo-dev <twin9774@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
uvicorn = "^0.30.6"
black = "^24.8.0"
python-multipart = "^0.0.10"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"
sqlalchemy = "^2.0.35"
pymysql = "^1.1.1"
cryptography = "^43.0.1"
aiomysql = "^0.2.0"
greenlet = "^3.1.1"
pytest = "^8.3.3"
sqlalchemy-utils = "^0.41.2"
alembic = "^1.13.3"
pydantic-settings = "^2.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
